requirements.txt
fastapi
uvicorn
requests
google-generativeai
python-jose
slowapi
beautifulsoup4